requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
flask
joblib
pyyaml
